+++ js/main.js
@@ -2,7 +2,7 @@
 import renderScreen from './render-screen';
 
 const questionData1 = {
-  type: `double-choose`,
+  type: `double`,
   text: `Угадайте для каждого изображения фото или рисунок?`,
   images: [
     {name: `question1`, rightValue: `paint`, src: `http://placehold.it/468x458`, alt: `Option 1`, width: `468`, height: `458`},
@@ -11,7 +11,7 @@ const questionData1 = {
 };
 
 const questionData2 = {
-  type: `choose`,
+  type: `single`,
   text: `Угадай, фото или рисунок?`,
   images: [
     {name: `question1`, rightValue: `paint`, src: `http://placehold.it/705x455`, alt: `Option 1`, width: `705`, height: `455`},
@@ -19,7 +19,7 @@ const questionData2 = {
 };
 
 const questionData3 = {
-  type: `finde`,
+  type: `triple`,
   text: `Найдите рисунок среди изображений`,
   images: [
     {name: `question1`, rightValue: false, src: `http://placehold.it/304x455`, alt: `Option 1`, width: `304`, height: `455`},
@@ -33,7 +33,7 @@ const initialState = {
   lives: 3,
   time: 0,
   answers: [],
-  questions: [questionData1, questionData3, questionData2, questionData1],
+  questions: [questionData1, questionData2, questionData3, questionData1],
 };
 
 renderScreen(initialState);
